requirements.txt
pytube
gtts
google-cloud-translate
google-cloud-texttospeech
srt
grpcio
urllib3
certifi
pydub
Flask